Created JHTS Restaurant Review classes
There are essentially three main classes: JHTSRestaurant, JHTSReview and JHTSUser (each subclassed from JHTSDatabaseObject).  My thought that even though JHTSReviews can be added and deleted, JHTSRestaurants and JHTSUser will remain permanently in the record (even if, say, a restaurant goes out of business, it is worth keeping its review history).  Since I figure that reading will happen an order of magnitude more often than modifications with new entries.  For this reason, the JHTSDatabaseObjects are stored in a growing vector of pointers, with each index essentially being its GUID. (Deleted JHTSReviews will have nils returned for those GUIDs.)

Each JHTSRestaurant will have properties of name, zip code, and a set of JHTSReview GUID’s associated with it.
Each JHTSReview has a star rating, review text, the GUID of the JHTSRestaurant it refers to, and the GUID of the JHTSUser posting it.
Each JHTSUser has properties of name, zip code, and a set of JHTSReview GUID’s that the user posted.

You can get the list of all JHTSRestaurants, list of JHTSRestaurants by a particular name (there may be one in different cities), a list of JHTSRestaurants in a certain mile range from a zip code, a list of JHTSRestaurants at or above a certain star rating, etc.  And of course a list of all JHTSReview associated with a JHTSRestaurants.

You can get a list of all JHTSUsers, JHTSUser by name or zip code, etc.  Also, a list of all JHTSReviews associated with a JHTSUser.  A JHTSUser can add a JHTSRestaurant, or add/delete/modify a JHTSReview.  The code checks to verify that any JHTSRestaurant is unique by name/zipcode (so that any “added” will not be repeated).  JHTSUsers may, however, share the same name and even zip code, there is no uniqueness there.  (However, you can see a list of JHTSUsers by name, to narrow down who you are looking for.)

An order operation is defined on JHTSDatabaseObject, so that an ordered set<> can be used in most of these cases for optimal performance.  For maximizing performance in reading data, a vector<> is used for the master lookup.

The code is written in straight C++, with the expectation that this would be used in a cross-platform sense (say, both on iPhone and Android).  So I used straight  C++ classes